Split bank admin / cashbook transaction api logic
So that it's clearer and different people can work on different files at the same time
