+++ mtp_api/apps/transaction/api/bank_admin/serializers.py
@@ -3,38 +3,8 @@ from django.db.models import Max
 
 from rest_framework import serializers
 
-from .signals import transaction_created
-from .models import Transaction
-
-
-class CashbookCreditedOnlyTransactionSerializer(serializers.ModelSerializer):
-    id = serializers.IntegerField(required=True)
-    credited = serializers.BooleanField(required=True)
-
-    class Meta:
-        model = Transaction
-        fields = (
-            'id',
-            'credited',
-        )
-
-
-class CashbookTransactionSerializer(serializers.ModelSerializer):
-    sender = serializers.CharField(source='sender_name')
-
-    class Meta:
-        model = Transaction
-        fields = (
-            'id',
-            'prisoner_number',
-            'amount',
-            'sender',
-            'received_at',
-            'prison',
-
-            'owner',
-            'credited',
-        )
+from transaction.signals import transaction_created
+from transaction.models import Transaction
 
 
 class BankAdminCreateTransactionListSerializer(serializers.ListSerializer):
