+++ mtp_api/apps/transaction/api/cashbook/permissions.py
