+++ mtp_api/apps/transaction/urls.py
@@ -2,15 +2,16 @@ from django.conf.urls import patterns, url, include
 
 from rest_framework import routers
 
-from .routers import CashbookTransactionRouter
-from . import views
+from .api.cashbook.routers import CashbookTransactionRouter
+from .api.cashbook import views as cashbook_views
+from .api.bank_admin import views as bank_admin_views
 
 
 cashbook_transaction_router = CashbookTransactionRouter()
-cashbook_transaction_router.register(r'transactions', views.CashbookTransactionView)
+cashbook_transaction_router.register(r'transactions', cashbook_views.CashbookTransactionView)
 
 admin_transaction_router = routers.DefaultRouter()
-admin_transaction_router.register(r'transactions', views.BankAdminTransactionView)
+admin_transaction_router.register(r'transactions', bank_admin_views.BankAdminTransactionView)
 
 urlpatterns = patterns('',
     url(r'^bank-admin/', include(admin_transaction_router.urls, namespace='bank-admin')),
